Refactored entire library into sensible files
