+++ taigit/src/libraries/GitHub/GitHub.js
@@ -1,16 +1,24 @@
-import {gethistoryPR} from "./src/getHistoryOfPR";
-
 export {getBranches, getNumBranchCommits} from "./src/branches";
-export {getCommitsInTimeWindow, getTotalCommits, getNumCommitsFromUser} from "./src/commitsInWindow";
-export {getNumOpenPullRequests} from "./src/numPullRequests";
-export {getNumComments} from "./src/avgComments";
-export {contributorData} from "./src/contributorData";
-export {authRedirect, getAuthToken} from "./src/auth"
-export {getContributerNames} from "./src/getUserList";
-export {usesMaven, usesGradle, usesCMake, usesMake, usesAnt, getBuilds} from "./src/findBuilds";
-export {getMemberInfo} from "./src/getMemberInfo";
-export {gethistoryPR} from "./src/getHistoryOfPR";
-export{getBytesOfCode} from "./src/bytesOfCode";
-export {getCodeAnalysis} from "./src/codeAnalysis";
-export {getUserRepos} from './src/reposAvailable';
-export {getUserInfo} from './src/getUserInfo';
+
+export {getCommitsInTimeWindow, getTotalCommits,
+    getNumCommitsFromUser} from "./src/commits";
+
+export {getHistoryPR, getNumComments,
+     getNumberCommentsPerPullRequest,
+     getNumClosedPullRequest,
+     getNumOpenPullRequests} from "./src/pullRequests";
+
+export {getContributorData, getContributorNames}from "./src/contributors";
+
+export {getMemberInfo} from "./src/members";
+
+export {getAuthRedirect, getAuthToken} from "./src/auth"
+
+export {usesMaven, usesGradle, usesCMake, usesMake, usesAnt, getBuilds}
+    from "./src/builds";
+
+export {getCodeAnalysis, getBytesOfCode}
+    from "./src/codeAnalysis";
+
+export {getUserInfo, getUserRepos}
+    from './src/users';
