+++ taigit/src/libraries/GitHub/src/codeAnalysis.ts
@@ -4,7 +4,7 @@ get cyclomatic complexity of python code
 */
 
 import axios from 'axios';
-import { counter } from '@fortawesome/fontawesome-svg-core'; 
+import { counter } from '@fortawesome/fontawesome-svg-core';
 
 /**
  *  check if the file is in the list of approved formats
@@ -12,9 +12,9 @@ import { counter } from '@fortawesome/fontawesome-svg-core';
  * @return bool true if the file is a java, python or c++ file
  */
 /*
-function 
+function
 verifyFormat(fileExt: string) : boolean{
-   let bool : boolean = false; 
+   let bool : boolean = false;
    if (fileExt === "java" || fileExt === "py" || fileExt === "cpp"){
        bool = true;
    }
@@ -48,12 +48,12 @@ getNumFiles(owner : string, repo : string, auth : string, path? : string) : Prom
     return fileCount;
 }
 /**
- * 
+ *
  * @param owner your user name on codacy
  * @param project your project name on codacy
  * @param ownerGit the github owner name--> in case owner of github repo is different from codacy owner
  * @param gitRepo the github repo name --> in case different from project name
- * @param auth Github auth key  
+ * @param auth Github auth key
  */
 
 export async function
@@ -62,7 +62,7 @@ getCodeAnalysis(owner: string, project:string, ownerGit : string, gitRepo : stri
    try{
     const response = await axios.get('https://taigit-auth.herokuapp.com/codacy/'+owner+"/"+project);
     jsonObj= response.data;
-    console.log("JSON data:\n",jsonObj); 
+    console.log("JSON data:\n",jsonObj);
     let fileCount : number = await getNumFiles(ownerGit, gitRepo, auth);
     jsonObj.fileCount = fileCount;
     console.log("JSONObj with fileCount\n:",jsonObj);
@@ -70,4 +70,27 @@ getCodeAnalysis(owner: string, project:string, ownerGit : string, gitRepo : stri
        console.log("Error1:\n", error);
    }
     return jsonObj;
+}
+
+/**
+ * Returns bytes of code of each programming language used in the
+ * github repo
+ * @param owner
+ * @param repo
+ * @param auth
+ */
+export async function
+getBytesOfCode(owner: string, repo : string, auth : string): Promise<any>{
+    try{
+        var config = {
+            headers: {'Authorization': "Bearer " + auth}
+        }
+        const getBytes = await axios.get('https://api.github.com/repos/'+owner +
+            '/' + repo + '/languages', config);
+
+        return getBytes.data;
+    }catch(error){
+        console.log("Error Bytes:\n", error);
+    }
+    return -1;
 }
