+++ taigit/src/libraries/GitHub/src/contributors.ts
@@ -1,22 +1,38 @@
-import axios from "axios";
+import axios from 'axios';
+
+export async function
+getContributorData(owner: string, repo: string, auth: string) {
+    try {
+        var config = {
+            headers: {'Authorization': "Bearer " + auth}
+        }
+        const contributorData = await axios.get("https://api.github.com/repos/" + owner +
+            "/" + repo + "/stats/contributors", config);
+        return contributorData.data;
+    } catch (error) {
+        console.log(error);
+    }
+    return -1;
+}
+
 /**
  * The following function returns the list of all contributer usernames in a
  * respository.
  * @param owner  The name of the owner of the repository in String format.
  * @param repo   The name of the Github repository in String format.
- * @return       The list consisting of usernames of All Contributers in the repo.
+ * @return       The list consisting of usernames of All Contributors in the repo.
  */
-export  async function
-getContributerNames(owner : string, repo: string, auth: string){
+export async function
+getContributorNames(owner : string, repo: string, auth: string){
     try{
         let contriNames: Array<string> = [];
         var config = {
             headers: {'Authorization': "Bearer " + auth}
         }
-        let listOfContributers = await axios.get("https://api.github.com/repos/" + owner +
+        let listOfContributors = await axios.get("https://api.github.com/repos/" + owner +
             "/" + repo + "/contributors", config);
-        if(listOfContributers.headers.hasOwnProperty("link")){
-            let last = listOfContributers.headers["link"].split(',');
+        if(listOfContributors.headers.hasOwnProperty("link")){
+            let last = listOfContributors.headers["link"].split(',');
             let total_pages_str = last[1]
             let total_pages = total_pages_str.substring(total_pages_str.lastIndexOf("page") + 5, total_pages_str.lastIndexOf(">"))
             total_pages = Number(total_pages)
@@ -31,7 +47,7 @@ getContributerNames(owner : string, repo: string, auth: string){
 
         }
         else {
-            listOfContributers.data.forEach(function (username: { login: string }) {
+            listOfContributors.data.forEach(function (username: { login: string }) {
                 contriNames.push(username.login)
             });
             return contriNames
@@ -41,7 +57,4 @@ getContributerNames(owner : string, repo: string, auth: string){
         console.log(error)
         return -1
     }
-
 }
-
-//getContributerNames("ser574-green-team", "taigit");
