+++ taigit/src/libraries/GitHub/src/users.ts
@@ -13,3 +13,18 @@ getUserRepos(owner: string, auth: string){
     }
     return -1;
 }
+
+export async function
+getUserInfo(auth: string){
+    try{
+        var config = {
+            headers: {'Authorization': "Bearer " + auth}
+        }
+        let userInfo = await axios.get("https://api.github.com/user", config);
+        return userInfo.data;
+    }
+    catch (error) {
+        console.log(error)
+        return -1
+    }
+}
